refactor(crane-mcp): split agent-launch.ts (531 LOC) into venture + engagement modules
- agent-launch.ts (~372 LOC): shared spawn infrastructure (buildChildEnv, spawnAgent,
  applyHermesTranslation, getStartupPrompt) + resolveAgent/validateAgentBinary/launchAgent
- engagement-launch.ts (~173 LOC): assertEngagementScope, fetchEngagementSecretsViaProxy,
  launchEngagement — SS engagement lifecycle with scope guard and proxy secret fetch
- launch-lib.ts barrel + main-router.ts import sites updated to new source locations

Two distinct lifecycles: venture (Infisical direct) vs SS engagement (crane-context proxy).
Verify clean. Gate: no files >500 LOC remain.
